Add User, Product logic

+++ src/main/java/dev/prom/delivery/service/OrderServiceImpl.java
@@ -1,12 +1,12 @@
 package dev.prom.delivery.service;
 
-import dev.prom.delivery.dto.OrderInputDto;
-import dev.prom.delivery.dto.OrderOutputDto;
+import dev.prom.delivery.exceptions.NotFoundException;
 import dev.prom.delivery.models.Order;
 import dev.prom.delivery.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 @Service
 public class OrderServiceImpl implements OrderService {
     private final OrderRepository orderRepository;
@@ -17,33 +17,31 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(Long id) {
-        orderRepository.deleteById(id);
+        if (orderRepository.existsById(id)) {
+            orderRepository.deleteById(id);
+        } else throw new NotFoundException("Order not found");
     }
 
     @Override
-    public OrderOutputDto updateOrder(Long id, OrderInputDto order) {
-        Order editedOrder = OrderMapper.INSTANCE.orderInputDtoToOrder(order);
-        editedOrder.setId(id);
-        Order updatedOrder = orderRepository.saveAndFlush(editedOrder);
-        return OrderMapper.INSTANCE.orderToOrderOutputDto(updatedOrder);
+    public Order updateOrder(Long id, Order order) {
+        if (orderRepository.existsById(id)) {
+            order.setId(id);
+        } else throw new NotFoundException("Order not found");
+        return orderRepository.saveAndFlush(order);
     }
 
     @Override
-    public OrderOutputDto createOrder(OrderInputDto order) {
-        Order newOrder = OrderMapper.INSTANCE.orderInputDtoToOrder(order);
-        newOrder = orderRepository.saveAndFlush(newOrder);
-        return OrderMapper.INSTANCE.orderToOrderOutputDto(newOrder);
+    public Order createOrder(Order order) {
+        return orderRepository.saveAndFlush(order);
     }
 
     @Override
-    public OrderOutputDto getOrderById(Long id) {
-        //обработать Optoinal на null
-        return OrderMapper.INSTANCE.orderToOrderOutputDto(orderRepository.findById(id).get());
+    public Order getOrderById(Long id) {
+        return orderRepository.findById(id).orElseThrow(() -> new NotFoundException("Order not found"));
     }
 
     @Override
-    public List<OrderOutputDto> getAllOrders() {
-        List<Order> orderList = orderRepository.findAll();
-        return OrderMapper.INSTANCE.ordersToOrderOutputDtos(orderList);
+    public List<Order> getAllOrders() {
+        return orderRepository.findAll();
     }
 }
